+++ test/gtc/gtc_integer.cpp
@@ -206,8 +206,8 @@ namespace uround
 
 		for(float f = 0.0f; f < 3.1f; f += 0.05f)
 		{
-			int RoundFast = glm::uround(f);
-			int RoundSTD = glm::round(f);
+			int RoundFast = static_cast<int>(glm::uround(f));
+			int RoundSTD = static_cast<int>(glm::round(f));
 			Error += RoundFast == RoundSTD ? 0 : 1;
 			assert(!Error);
 		}
